Change Colour Mode
Change to HSL from HSB as you can get HSL values from the browsers colour
pickers and are CSS/HTML compatable colour Schemes. HSB is purely a p5.js
thing.

Create an object to hold all the colors we're going to use in one place to
make them easier to read and use around the codebase. Great example of how
an objects key:value creates meaning for the programmer.